Update message_memory.py : Migrated Message Functions
